Make BiMultiMap package private.
