+++ library/src/com/google/maps/android/geojson/GeoJsonRenderer.java
@@ -7,7 +7,6 @@ import com.google.android.gms.maps.model.Polygon;
 import com.google.android.gms.maps.model.PolygonOptions;
 import com.google.android.gms.maps.model.Polyline;
 import com.google.android.gms.maps.model.PolylineOptions;
-import com.google.maps.android.BiMultiMap;
 
 import java.util.ArrayList;
 import java.util.HashMap;
